Move empty project to dedicated card
