+++ web-common/src/features/welcome/TitleContent.svelte
@@ -1,36 +1,14 @@
 <script lang="ts">
-  import { goto } from "$app/navigation";
   import Add from "@rilldata/web-common/components/icons/Add.svelte";
   import RillLogoSquareNegative from "@rilldata/web-common/components/icons/RillLogoSquareNegative.svelte";
   import RadixH1 from "@rilldata/web-common/components/typography/RadixH1.svelte";
   import Subheading from "@rilldata/web-common/components/typography/Subheading.svelte";
   import AddSourceModal from "@rilldata/web-common/features/sources/add-source/AddSourceModal.svelte";
-  import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
-  import { createRuntimeServiceUnpackEmpty } from "../../runtime-client";
-  import { EMPTY_PROJECT_TITLE } from "./constants";
 
   let showAddSourceModal = false;
   const openShowAddSourceModal = () => {
     showAddSourceModal = true;
   };
-
-  const unpackEmptyProject = createRuntimeServiceUnpackEmpty();
-  async function startWithEmptyProject() {
-    $unpackEmptyProject.mutate(
-      {
-        instanceId: $runtime.instanceId,
-        data: {
-          title: EMPTY_PROJECT_TITLE,
-          force: true,
-        },
-      },
-      {
-        onSuccess: () => {
-          goto("/");
-        },
-      }
-    );
-  }
 </script>
 
 <section class="flex flex-col gap-y-6 items-center text-center">
@@ -60,12 +38,6 @@
       <Add className="text-white" />
       Add data
     </div>
-  </button>
-  <button
-    class="px-2 font-medium text-xs text-blue-500 hover:text-blue-400"
-    on:click={startWithEmptyProject}
-  >
-    Or start with an empty project
   </button>
   {#if showAddSourceModal}
     <AddSourceModal
